moved functions operating on Nodes to network_components.py and changed tests accordingly
