+++ tensornetwork/contractors/opt_einsum_paths/utils.py
@@ -13,10 +13,9 @@
 # limitations under the License.
 """Helper methods for `path_contractors`."""
 
-from tensornetwork import network
-from tensornetwork import network_components
+from tensornetwork.network import TensorNetwork
+from tensornetwork.network_components import BaseNode
 from typing import Any, Callable, Dict, List, Set, Tuple
-
 # `opt_einsum` algorithm method typing
 Algorithm = Callable[[List[Set[int]], Set[int], Dict[int, int]], List[
     Tuple[int, int]]]
@@ -27,9 +26,8 @@ def multi_remove(elems: List[Any], indices: List[int]) -> List[Any]:
   return [i for j, i in enumerate(elems) if j not in indices]
 
 
-def get_path(
-    net: network.TensorNetwork, algorithm: Algorithm
-) -> Tuple[List[Tuple[int, int]], List[network_components.BaseNode]]:
+def get_path(net: TensorNetwork, algorithm: Algorithm
+            ) -> Tuple[List[Tuple[int, int]], List[BaseNode]]:
   """Calculates the contraction paths using `opt_einsum` methods.
 
   Args:
